Add TS ServiceModel Logger helmet
