--- startup/cors.js
@@ -1,5 +1,0 @@
-const cors = require("cors");
-
-module.exports = function(app) {
-	app.use(cors());
-};
