+++ startup/cors.ts
@@ -1,0 +1,7 @@
+import { Express } from "express";
+
+const cors = require("cors");
+
+module.exports = function(app: Express) {
+	app.use(cors());
+};
