--- startup/db.js
@@ -1,9 +1,0 @@
-const winston = require('winston');
-const mongoose = require('mongoose');
-const config = require('config');
-
-module.exports = function() {
-	const db = config.get('db');
-	mongoose.connect(db)
-			.then(() => winston.info(`Connected to ${db}...`));
-}
