+++ startup/db.ts
@@ -1,0 +1,9 @@
+import winston from 'winston';
+import mongoose from 'mongoose';
+import config from 'config';
+
+module.exports = function() {
+	const db: string = config.get('db');
+	mongoose.connect(db)
+			.then(() => winston.info(`Connected to ${db}...`));
+}
